feat: dto 유효성 검증 및 Bad Request 예외 처리 구현
