+++ build.gradle
@@ -32,6 +32,7 @@ dependencies {
     /* Boot */
     implementation 'org.springframework.boot:spring-boot-starter-web:3.2.4'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa:3.2.4'
+    implementation 'org.springframework.boot:spring-boot-starter-validation:3.2.4'
     developmentOnly 'org.springframework.boot:spring-boot-devtools:3.2.4'
     annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor:3.2.4'
     testImplementation 'org.springframework.boot:spring-boot-starter-test:3.2.4'
@@ -44,6 +45,9 @@ dependencies {
     runtimeOnly 'com.h2database:h2:2.2.224'
     runtimeOnly 'com.mysql:mysql-connector-j:8.3.0'
 
+    /* Redis */
+    implementation 'org.springframework.boot:spring-boot-starter-data-redis:3.2.4'
+
     /* Security */
     implementation 'org.springframework.boot:spring-boot-starter-security:3.2.4'
     testImplementation 'org.springframework.security:spring-security-test:6.2.3'
@@ -52,7 +56,7 @@ dependencies {
     implementation 'io.jsonwebtoken:jjwt-api:0.12.5'
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.5'
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.5'
-    
+
     /* Spring Rest Docs */
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc:3.0.1'
     testImplementation "com.epages:restdocs-api-spec-mockmvc:0.18.2"
